+++ resources/js/project/components/Team.vue
@@ -87,9 +87,11 @@ export default {
             })
         },
         deleteTeam(id) {
-            this.$axios.delete('api/team/'+id).then((res) => {
-                this.listPage();
-            })
+            if (confirm("¿Estás seguro de que deseas eliminar el registro?")) {
+                this.$axios.delete('api/team/'+id).then((res) => {
+                    this.listPage();
+                })
+            }
         }
     },
     async mounted() {
